+++ package.json
@@ -32,7 +32,6 @@
     }
   },
   "dependencies": {
-    "just-observable": "^0.3.0",
     "make-list-provider": "^0.2.2"
   },
   "devDependencies": {
