+++ src/global.d.ts
@@ -1,0 +1,9 @@
+type HTMLDivProps = React.DetailedHTMLProps<
+	React.HTMLAttributes<HTMLDivElement>,
+	HTMLDivElement
+>;
+
+type HTMLCanvasProps = React.DetailedHTMLProps<
+	React.HTMLAttributes<HTMLCanvasElement>,
+	HTMLCanvasElement
+>;
